Ease selection of contents of delimiters
This change makes it easier to select the contents inside of
delimiters. When seeking where to put the cursor for a selection, the
cursor will be placed inside the delimiters if the seek location is on
the delimiter. If the cursor is to the left of the left delimiter or
the right of the right delimiter, then the whole of the delimiter
command will be selected.

This does not affect keyboard-based selection.