Add prevent spaces from behaving like tabs in root blocks.
This commit adds another configuration parameter for the spaceBehavesLikeTab
option.  Previously, the only valid values were 'true' and 'false'.  Where
'false' simply inserted spaces and 'true' made spaces behave like tab.  It is
now possible to add a string 'exceptRootBlock'.  'exceptRootBlock' behaves
the same as 'true' unless the cursor is in the root block, where it will
insert a space instead.

This feature allows the entry of mixed fractions in the root block, which
'true' was unable to do.

This feature is purposfully overloading the semantics of 'spaceBehavesLikeTab'
to add a new configuration option because this causes the smallest divergeance
from the MathQuill root project.  All things being equal, however, the better
change would be to rename the configuration option to 'spaceBehavior' and use
strings for all options: 'default', 'asSpaces', 'asTabs', and
'asTabsExceptForRootBlock'.  Where 'default' and 'asSpaces' are currently the
same.