Fix unit tests for selection
The strictOperator selection pieces require that Cursor have a valid
options.  This simply means that we have to pass an options object into
the constructor.  The unit test previously left the constructor empty,
so this is a simple fix.  Also added a prayer for a valid object in
Cursor to make it easier to find this type of error in the future.

Since it does not seem to be the approach of MathQuill to use default
values in constructor, this was avoided.  This could have also been
solved by:
this.options = options || {};