Centralize LESS user overrides
Created a variables.less to house all of the variables and
default-guarded mixins that users may wish to override.

Pulled the current illegal CSS values in the variables file as
variables.  The final output CSS contained several instance of
HighlightText and other illegal strings.  While these are dutifully
ignored by browsers it seems the intent was to make these variables.

I also created a default-guarded mixin for styling empty blocks.  The
mathquill default staying of a gray box is now easily changeable.  It
is also easier to share this box to other parts of the mathquill CSS.

The only effective change to the output file, mathquill.css, is the
removal of the redundant background transparency for empty root blocks
and a now duplicated border color as Highlight has become a real
variable.  There was a TODO here as well.  It’s now TO-DONE.